requirements.txt
pygame==2.5.0
python-socketio==5.8.0
fastapi==0.100.0
sqlalchemy==2.0.0
pytest==7.4.0
black==23.3.0
python-dotenv==1.0.0
pydantic==2.0.0
ruff==0.1.6
uvicorn==0.24.0
pytest-cov==4.1.0
mypy==1.7.0
pygame-gui==0.6.9 